Move ClientImagesList back into DreamList cause idk what I was thinking
